requirements.txt
dash
numpy
pandas
plotly
pyyaml
requests
